% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyR.R
\name{nav_match}
\alias{nav_match}
\title{Match events to navigation data.}
\usage{
nav_match(event.date, nav.df)
}
\arguments{
\item{event.date}{Event date, in POSIXct format.}

\item{nav.df}{Data frame containing navigation data from Winfrog.}
}
\value{
Data frame containing the nearest `nav_id` in time and the time difference (`lag_s`) in seconds.
}
\description{
Match events to navigation data.
}
